Added --min_tier: specify the minimum tier
ProtonVPN advertises certain features exclusive to their higher-tier
services (e.g., https://protonvpn.com/support/watch-netflix-with-vpn/),
but how does one actually use such features? Connecting to the "fastest"
server doesn't really work, because this often results in a "Basic"
server. One could manually select the specific server, but this is not
a good UI. The --min_tier option allows one to select the fastest
server, with a tier greater than or equal to --min_tier, providing
that the user's plan is compatible with the given --min_tier.

--min_tier is applied every place it could be relevant: fastest server,
fastest server in a country, random server, and p2p server. There is
currently no possibility of using a lower tier server in combination
with TOR or Secure Core, so it is not applied in combination with those
features.