:memo: docs: Add install instructions for Fedora/CentOS/RHEL
As of this morning, `protonvpn-cli` is available in the official Fedora
and EPEL repositories. It is possible to directly install the package
from the official repositories for these distributions. This simplifies
the installation instructions for users on these distributions and also
allows users to use their package managers to obtain future updates.

The README and USAGE docs are expanded and table-of-contents slightly
modified to include this section. I also moved the _Initialization_ step
farther down on USAGE since you have to follow that step regardless of
how you set up `protonvpn-cli`.

Cheers!

Signed-off-by: Justin W. Flory <git@jwf.io>